Fix compile error on Linux and Mac

+++ benchmark/benchmark.cpp
@@ -7,8 +7,8 @@ static void BM_RenderRandomScene(benchmark::State& state) {
 
     for (auto _ : state) {
 
-        const int width = 160;
-        const int height = 120;
+        const size_t width = 160;
+        const size_t height = 120;
         const int numRaysPerPixel = 100;
 
         auto world = World::randomScene();
